refactor

+++ 08_encapsulation_exercise/01_wild_cat_zoo/project/animal.py
@@ -1,5 +1,5 @@
 class Animal:
-    def __init__(self, name: str, gender: str, age: int, money_for_care: int):
+    def __init__(self, name: str, gender: str, age: int, money_for_care: float):
         self.name = name
         self.gender = gender
         self.age = age
