+++ 08_encapsulation_exercise/01_wild_cat_zoo/project/worker.py
@@ -1,5 +1,5 @@
 class Worker:
-    def __init__(self, name: str, age: int, salary: int):
+    def __init__(self, name: str, age: int, salary: float):
         self.name = name
         self.age = age
         self.salary = salary
